requirements.txt
pillow
streamlit
pytesseract
opencv-python
numpy
